organize di modules and fix missing adapter
