+++ backintime-debugger/data/src/jvmMain/kotlin/com/github/kitakkun/backintime/debugger/data/driver/SqlDriver.jvm.kt
@@ -1,4 +1,4 @@
-package com.github.kitakkun.backintime.debugger.data
+package com.github.kitakkun.backintime.debugger.data.driver
 
 import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
